move pcc stabilization
